requirements.txt
Flask==1.1.1
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==19.9.0
google-cloud-vision==0.39.0
google-cloud-storage==1.19.1
Wand==0.5.9
pdfminer.six